requirements.txt
gym-minigrid==1.1.0
gymnasium==0.29.1
numpy>=1.24,<2
stable-baselines3>=2.3.0
